agentic-backend/requirements.txt
google-adk
fastapi
uvicorn[standard]
python-dotenv